requirements.txt
gdown
torch
torchvision
Pillow
PySimpleGUI